requirements.txt
azure-core==1.26.2
azure-cosmos==4.3.0
pytest==6.2.5